Style: 네이버 뉴스 검색 함수 주석 추가
